pyproject.toml
[tool.poetry]
name = "flowgpt-multi-platform-bot"
version = "1.0.0"
description = "Multi-platform chatbot integration for FlowGPT - supports Telegram, Discord, Slack, and LINE"
authors = ["Devin AI <devin@cognition.ai>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
httpx = "^0.27.0"
python-dotenv = "^1.0.0"

# Telegram
python-telegram-bot = {version = "^21.0", optional = true}

# Discord
"discord.py" = {version = "^2.3.0", optional = true}

# Slack
slack-bolt = {version = "^1.18.0", optional = true}

# LINE
line-bot-sdk = {version = "^3.5.0", optional = true}

# API Server (for LINE webhooks)
fastapi = {version = "^0.109.0", optional = true}
uvicorn = {version = "^0.27.0", optional = true}

[tool.poetry.extras]
telegram = ["python-telegram-bot"]
discord = ["discord.py"]
slack = ["slack-bolt"]
line = ["line-bot-sdk", "fastapi", "uvicorn"]
all = ["python-telegram-bot", "discord.py", "slack-bolt", "line-bot-sdk", "fastapi", "uvicorn"]

[tool.poetry.scripts]
flowgpt-bot = "src.main:run"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
black = "^24.0.0"
ruff = "^0.2.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
